refactor: fast-usdc files to own package
