+++ packages/fast-usdc/tools/cosmoverse-mock.js
