+++ packages/fast-usdc/tools/eth-mock.ts
